requirements.txt
streamlit
pandas
plotly 
streamlit-aggrid
snowflake-connector-python